requirements.txt
Django==2.0
pytz==2017.3
python-dateutil==2.6.1
requests==2.13.0
beautifulsoup4==4.6.0
djangorestframework==3.7.7
django-rest-swagger==2.1.2
lxml==4.1.1
psycopg2==2.5.4